Move flow main() to direktiv main().
