+++ cmd/flow/tests-instance-common.go
@@ -1,4 +1,4 @@
-package main
+package flow
 
 /*
