+++ cmd/flow/tests-instance-retries.go
@@ -1,4 +1,4 @@
-package main
+package flow
 
 /*
