+++ cmd/flow/tests-vars.go
@@ -1,4 +1,4 @@
-package main
+package flow
 
 /*
